Update cookie handling and security configuration
This commit extends the max age of cookies and improves cookie value retrieval in `CookieUtility`. A new `BearerTokenConfiguration` class was introduced and integrated into `SecurityConfiguration`. In `AuthenticationService`, the 'remember me' logic was revised and integrated with the `RefreshTokenService` and `JwtService` where token expiry times were extended. Permissions for `/ai/**` endpoints have been updated to require 'ADMIN' role. The `RefreshTokenRepository` now includes a method to get tokens by `ApplicationUser`.

+++ src/main/java/com/conner/assistant/services/RefreshTokenService.java
@@ -23,8 +23,7 @@ public class RefreshTokenService {
         RefreshToken refreshToken = RefreshToken.builder()
                 .applicationUser(userRepository.findByUsername(username).orElseThrow())
                 .token(UUID.randomUUID().toString())
-                // set expiry of refresh token in milliseconds
-                .expiryDate(Instant.now().plusMillis(30000))
+                .expiryDate(Instant.now().plusSeconds(60*60))
                 .build();
         return refreshTokenRepository.save(refreshToken);
     }
@@ -39,8 +38,8 @@ public class RefreshTokenService {
     }
 
     public RefreshToken verifyRefreshToken(String refreshToken){
-        RefreshToken token = findByToken(refreshToken).orElseThrow();
-        if(token.getExpiryDate().compareTo(Instant.now())<0){
+        RefreshToken token = findByToken(refreshToken).orElse(null);
+        if(token != null && token.getExpiryDate().compareTo(Instant.now())<0){
             refreshTokenRepository.delete(token);
             throw new RuntimeException("Refresh token is expired or not valid. Please make a new login..!");
         }
